Fix error where a class object name can not to implicit conversion to string
When a custom tag called "owner" is passed in via CLI, eg. `yard doc --tag owner:"owner"` classes with comments would fail generating their doc page.

This obviously happens because of some sort of collision, but what the collision is on is a whole other story.

This fixes that issue by explicitly calling `.to_s` on `value` and changing that class object to a string.