Return only the path if creating docs for a single library
We would run into an error where version_supplied is not definited in Yard::Server::DocServerHelper#base_path on router.request. This probably happens because it's refined/overloaded but the request object we get in that method doesn't include that definition. We are using a single library anyways and don't need that check anyways, so let's skip it.